重构数据库默认值
